Relocate list secrets query to a separate file
